Admin can update plan

+++ database/migrations/2017_03_01_015927_create_plan_level_sponsors_table.php
@@ -18,7 +18,7 @@ class CreatePlanLevelSponsorsTable extends Migration
             $table->integer('plan_id')->unsigned();
             $table->integer('level');
             $table->char('type', 1);
-            $table->decimal('commission');
+            $table->double('commission');
         });
     }
